Modify Add function.

+++ 100DaysOfCodeProjects/calculator-100days/calculator.go
@@ -1,5 +1,5 @@
 package calculator
 
-func Add(x int, y int) int {
+func Add(x float64, y float64) float64 {
 	return x + y
 }
